+++ frontend/src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginPage from "@/components/Login"
+import LoginPage from "@/app/auth/login/Login"
 
 export default function page(){
     return(
