+++ frontend/src/app/group/create/CreateGroup.tsx
