+++ frontend/src/app/members/add/[...groupID]/page.tsx
@@ -1,4 +1,4 @@
-import AddMembers from '@/components/AddMembers'
+import AddMembers from '@/app/members/add/[...groupID]/AddMembers'
 import React from 'react'
 
 export default function Page() {
